refactor: extract codegen/dce.rs — dead code elimination and topo sort
Move dead_code_eliminate, topo_sort_definitions, and collect_calls_in_expr
into a dedicated dce module. Pure mechanical extraction.
